+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 * Update: Citation info now points to `SciPost paper <https://scipost.org/SciPostPhys.7.6.080>`_ (`#61`_)
 * Added: parameters `col_formats` and `col_headers` to customize the output of ``krotov.info_hooks.print_table`` (`#65`_)
 * Added: info-hooks now have access to the additional arguments `propagator`, `chi_constructor`, `mu`, `sigma`, `iter_start`, and `iter_stop` (`#66`_)
+* Added: parameter `keep_original_objectives` to ``krotov.objectives.ensemble_objectives`` (`#67`_)
 
 0.5.0 (2019-12-04)
 ------------------
@@ -107,3 +108,4 @@ History
 .. _#61: https://github.com/qucontrol/krotov/issues/61
 .. _#65: https://github.com/qucontrol/krotov/issues/65
 .. _#66: https://github.com/qucontrol/krotov/issues/66
+.. _#67: https://github.com/qucontrol/krotov/issues/67